snmp-ups: fix loss of precision when setting values
su_setvar() was losing precision when converting and casting the
provided values to send to the SNMP agent. As an example, with an OID in
millivolt (multiplier set to 0.001), when providing 238 (V) using upsrw,
the value sent to the SNMP agent was 237999, so leaking 0.1 volt

+++ drivers/snmp-ups.h
@@ -117,7 +117,7 @@ typedef struct {
 typedef struct {
 	const char   *info_type;	/* INFO_ or CMD_ element */
 	int           info_flags;	/* flags to set in addinfo */
-	float         info_len;		/* length of strings if STR,
+	double        info_len;		/* length of strings if STR,
 								 * cmd value if CMD, multiplier otherwise. */
 	const char   *OID;			/* SNMP OID or NULL */
 	const char   *dfl;			/* default value */
